removed old files, and renamed things
